transactor/copycat: refactor Server, Serializers out of Copycat
